*: remove MetricsRegistryTrait
