+++ core/src/polling_monitor/metrics.rs
@@ -1,8 +1,7 @@
 use std::sync::Arc;
 
 use graph::{
-    components::metrics::MetricsRegistryTrait,
-    prelude::DeploymentHash,
+    prelude::{DeploymentHash, MetricsRegistry},
     prometheus::{Counter, Gauge},
 };
 
@@ -14,7 +13,7 @@ pub struct PollingMonitorMetrics {
 }
 
 impl PollingMonitorMetrics {
-    pub fn new(registry: Arc<dyn MetricsRegistryTrait>, subgraph_hash: &DeploymentHash) -> Self {
+    pub fn new(registry: Arc<MetricsRegistry>, subgraph_hash: &DeploymentHash) -> Self {
         let requests = registry
             .new_deployment_counter(
                 "polling_monitor_requests",
