+++ graph/src/blockchain/firehose_block_stream.rs
@@ -3,7 +3,6 @@ use super::client::ChainClient;
 use super::{Blockchain, TriggersAdapter};
 use crate::blockchain::block_stream::FirehoseCursor;
 use crate::blockchain::TriggerFilter;
-use crate::components::metrics::MetricsRegistryTrait;
 use crate::prelude::*;
 use crate::util::backoff::ExponentialBackoff;
 use crate::{firehose, firehose::FirehoseEndpoint};
@@ -23,7 +22,7 @@ struct FirehoseBlockStreamMetrics {
 }
 
 impl FirehoseBlockStreamMetrics {
-    pub fn new(registry: Arc<dyn MetricsRegistryTrait>, deployment: DeploymentHash) -> Self {
+    pub fn new(registry: Arc<MetricsRegistry>, deployment: DeploymentHash) -> Self {
         Self {
             deployment,
 
@@ -116,7 +115,7 @@ where
         filter: Arc<C::TriggerFilter>,
         start_blocks: Vec<BlockNumber>,
         logger: Logger,
-        registry: Arc<dyn MetricsRegistryTrait>,
+        registry: Arc<MetricsRegistry>,
     ) -> Self
     where
         F: FirehoseMapper<C> + 'static,
