+++ graph/src/components/metrics/subgraph.rs
@@ -6,7 +6,7 @@ use std::collections::HashMap;
 use std::sync::Arc;
 
 use super::stopwatch::StopwatchMetrics;
-use super::MetricsRegistryTrait;
+use super::MetricsRegistry;
 
 pub struct SubgraphInstanceMetrics {
     pub block_trigger_count: Box<Histogram>,
@@ -20,7 +20,7 @@ pub struct SubgraphInstanceMetrics {
 
 impl SubgraphInstanceMetrics {
     pub fn new(
-        registry: Arc<dyn MetricsRegistryTrait>,
+        registry: Arc<MetricsRegistry>,
         subgraph_hash: &str,
         stopwatch: StopwatchMetrics,
     ) -> Self {
@@ -79,7 +79,7 @@ impl SubgraphInstanceMetrics {
         self.trigger_processing_duration.observe(duration);
     }
 
-    pub fn unregister(&self, registry: Arc<dyn MetricsRegistryTrait>) {
+    pub fn unregister(&self, registry: Arc<MetricsRegistry>) {
         registry.unregister(self.block_processing_duration.clone());
         registry.unregister(self.block_trigger_count.clone());
         registry.unregister(self.trigger_processing_duration.clone());
@@ -94,7 +94,7 @@ pub struct SubgraphCountMetric {
 }
 
 impl SubgraphCountMetric {
-    pub fn new(registry: Arc<dyn MetricsRegistryTrait>) -> Self {
+    pub fn new(registry: Arc<MetricsRegistry>) -> Self {
         let running_count = registry
             .new_gauge(
                 "deployment_running_count",
