+++ graphql/src/metrics.rs
@@ -3,9 +3,8 @@ use std::fmt;
 use std::sync::Arc;
 use std::time::Duration;
 
-use graph::components::metrics::MetricsRegistryTrait;
 use graph::data::query::QueryResults;
-use graph::prelude::{DeploymentHash, GraphQLMetrics as GraphQLMetricsTrait};
+use graph::prelude::{DeploymentHash, GraphQLMetrics as GraphQLMetricsTrait, MetricsRegistry};
 use graph::prometheus::{CounterVec, Gauge, Histogram, HistogramVec};
 
 pub struct GraphQLMetrics {
@@ -77,7 +76,7 @@ impl GraphQLMetricsTrait for GraphQLMetrics {
 }
 
 impl GraphQLMetrics {
-    pub fn new(registry: Arc<dyn MetricsRegistryTrait>) -> Self {
+    pub fn new(registry: Arc<MetricsRegistry>) -> Self {
         let query_execution_time = registry
             .new_histogram_vec(
                 "query_execution_time",
@@ -141,7 +140,7 @@ impl GraphQLMetrics {
 
     // Tests need to construct one of these, but normal code doesn't
     #[cfg(debug_assertions)]
-    pub fn make(registry: Arc<dyn MetricsRegistryTrait>) -> Self {
+    pub fn make(registry: Arc<MetricsRegistry>) -> Self {
         Self::new(registry)
     }
