+++ node/src/store_builder.rs
@@ -3,8 +3,7 @@ use std::{collections::HashMap, sync::Arc};
 
 use futures::future::join_all;
 use graph::blockchain::ChainIdentifier;
-use graph::components::metrics::MetricsRegistryTrait;
-use graph::prelude::{o, NodeId};
+use graph::prelude::{o, MetricsRegistry, NodeId};
 use graph::url::Url;
 use graph::{
     prelude::{info, CheapClone, Logger},
@@ -30,7 +29,7 @@ pub struct StoreBuilder {
     /// Map network names to the shards where they are/should be stored
     chains: HashMap<String, ShardName>,
     pub coord: Arc<PoolCoordinator>,
-    registry: Arc<dyn MetricsRegistryTrait>,
+    registry: Arc<MetricsRegistry>,
 }
 
 impl StoreBuilder {
@@ -42,7 +41,7 @@ impl StoreBuilder {
         node: &NodeId,
         config: &Config,
         fork_base: Option<Url>,
-        registry: Arc<dyn MetricsRegistryTrait>,
+        registry: Arc<MetricsRegistry>,
     ) -> Self {
         let primary_shard = config.primary_store().clone();
 
@@ -98,7 +97,7 @@ impl StoreBuilder {
         node: &NodeId,
         config: &Config,
         fork_base: Option<Url>,
-        registry: Arc<dyn MetricsRegistryTrait>,
+        registry: Arc<MetricsRegistry>,
     ) -> (
         Arc<SubgraphStore>,
         HashMap<ShardName, ConnectionPool>,
@@ -168,7 +167,7 @@ impl StoreBuilder {
         subgraph_store: Arc<SubgraphStore>,
         chains: HashMap<String, ShardName>,
         networks: Vec<(String, Vec<ChainIdentifier>)>,
-        registry: Arc<dyn MetricsRegistryTrait>,
+        registry: Arc<MetricsRegistry>,
     ) -> Arc<DieselStore> {
         let networks = networks
             .into_iter()
@@ -205,7 +204,7 @@ impl StoreBuilder {
         node: &NodeId,
         name: &str,
         shard: &Shard,
-        registry: Arc<dyn MetricsRegistryTrait>,
+        registry: Arc<MetricsRegistry>,
         coord: Arc<PoolCoordinator>,
     ) -> ConnectionPool {
         let logger = logger.new(o!("pool" => "main"));
@@ -241,7 +240,7 @@ impl StoreBuilder {
         node: &NodeId,
         name: &str,
         shard: &Shard,
-        registry: Arc<dyn MetricsRegistryTrait>,
+        registry: Arc<MetricsRegistry>,
         coord: Arc<PoolCoordinator>,
     ) -> (Vec<ConnectionPool>, Vec<usize>) {
         let mut weights: Vec<_> = vec![shard.weight];
