+++ store/postgres/src/connection_pool.rs
@@ -7,10 +7,9 @@ use diesel::{
 use diesel::{sql_query, RunQueryDsl};
 
 use graph::cheap_clone::CheapClone;
-use graph::components::metrics::MetricsRegistryTrait;
 use graph::constraint_violation;
-use graph::prelude::tokio;
 use graph::prelude::tokio::time::Instant;
+use graph::prelude::{tokio, MetricsRegistry};
 use graph::slog::warn;
 use graph::util::timed_rw_lock::TimedMutex;
 use graph::{
@@ -310,7 +309,7 @@ impl ConnectionPool {
         pool_size: u32,
         fdw_pool_size: Option<u32>,
         logger: &Logger,
-        registry: Arc<dyn MetricsRegistryTrait>,
+        registry: Arc<MetricsRegistry>,
         coord: Arc<PoolCoordinator>,
     ) -> ConnectionPool {
         let state_tracker = PoolStateTracker::new();
@@ -593,7 +592,7 @@ struct EventHandler {
 impl EventHandler {
     fn new(
         logger: Logger,
-        registry: Arc<dyn MetricsRegistryTrait>,
+        registry: Arc<MetricsRegistry>,
         wait_stats: PoolWaitStats,
         const_labels: HashMap<String, String>,
         state_tracker: PoolStateTracker,
@@ -711,7 +710,7 @@ impl PoolInner {
         pool_size: u32,
         fdw_pool_size: Option<u32>,
         logger: &Logger,
-        registry: Arc<dyn MetricsRegistryTrait>,
+        registry: Arc<MetricsRegistry>,
         state_tracker: PoolStateTracker,
     ) -> PoolInner {
         let logger_store = logger.new(o!("component" => "Store"));
@@ -1157,7 +1156,7 @@ impl PoolCoordinator {
         postgres_url: String,
         pool_size: u32,
         fdw_pool_size: Option<u32>,
-        registry: Arc<dyn MetricsRegistryTrait>,
+        registry: Arc<MetricsRegistry>,
     ) -> ConnectionPool {
         let is_writable = !pool_name.is_replica();
