+++ store/postgres/src/writable.rs
@@ -5,13 +5,13 @@ use std::time::Duration;
 use std::{collections::BTreeMap, sync::Arc};
 
 use graph::blockchain::block_stream::FirehoseCursor;
-use graph::components::metrics::MetricsRegistryTrait;
 use graph::components::store::ReadStore;
 use graph::components::store::{DeploymentCursorTracker, EntityKey};
 use graph::data::subgraph::schema;
 use graph::data_source::CausalityRegion;
 use graph::prelude::{
-    BlockNumber, Entity, Schema, SubgraphDeploymentEntity, SubgraphStore as _, BLOCK_NUMBER_MAX,
+    BlockNumber, Entity, MetricsRegistry, Schema, SubgraphDeploymentEntity, SubgraphStore as _,
+    BLOCK_NUMBER_MAX,
 };
 use graph::slog::info;
 use graph::util::bounded_queue::BoundedQueue;
@@ -586,7 +586,7 @@ impl Queue {
         logger: Logger,
         store: Arc<SyncStore>,
         capacity: usize,
-        registry: Arc<dyn MetricsRegistryTrait>,
+        registry: Arc<MetricsRegistry>,
     ) -> Arc<Self> {
         async fn start_writer(queue: Arc<Queue>, logger: Logger) {
             loop {
@@ -855,7 +855,7 @@ impl Writer {
         logger: Logger,
         store: Arc<SyncStore>,
         capacity: usize,
-        registry: Arc<dyn MetricsRegistryTrait>,
+        registry: Arc<MetricsRegistry>,
     ) -> Self {
         info!(logger, "Starting subgraph writer"; "queue_size" => capacity);
         if capacity == 0 {
@@ -987,7 +987,7 @@ impl WritableStore {
         subgraph_store: SubgraphStore,
         logger: Logger,
         site: Arc<Site>,
-        registry: Arc<dyn MetricsRegistryTrait>,
+        registry: Arc<MetricsRegistry>,
     ) -> Result<Self, StoreError> {
         let store = Arc::new(SyncStore::new(subgraph_store, logger.clone(), site)?);
         let block_ptr = Mutex::new(store.block_ptr().await?);
